+++ build.config.ts
@@ -10,6 +10,7 @@ export default defineBuildConfig({
     'src/runtime/server/api/callback.get',
     'src/runtime/server/api/session.get',
     'src/runtime/server/api/logout.post',
+    'src/runtime/server/api/account-authorize.get',
     'src/runtime/client/routes',
     'src/runtime/client/session',
     'src/runtime/client/middleware',